moved simpleTextParser to io.ebean.text
